Model file renames and import fixes
