+++ st2common/st2common/models/api/execution.py
@@ -19,7 +19,7 @@ import six
 
 from st2common.util import isotime
 from st2common.models.api.base import BaseAPI
-from st2common.models.db.history import ActionExecutionDB
+from st2common.models.db.execution import ActionExecutionDB
 from st2common.models.api.reactor import TriggerTypeAPI, TriggerAPI, TriggerInstanceAPI
 from st2common.models.api.rule import RuleAPI
 from st2common.models.api.action import RunnerTypeAPI, ActionAPI, LiveActionAPI
